[CHORE] # 11 - PopUpController 상속 받게 수정
