+++ dnd-11th-4-iOS/dnd-11th-4-iOS/Presetation/PopUp/PopupViewController.swift
@@ -9,7 +9,7 @@ import UIKit
 import SnapKit
 
 class PopUpViewController: UIViewController {
-    private let popUpView = MDPopUpView()
+    
     private let dimmedView = UIView()
     
     init() {
@@ -25,8 +25,8 @@ class PopUpViewController: UIViewController {
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         
-        dimmedView.backgroundColor = .gray10
-        dimmedView.alpha = 1
+        dimmedView.backgroundColor = .black2
+        dimmedView.alpha = 0
         self.view.addSubview(dimmedView)
         
         dimmedView.snp.makeConstraints { make in
@@ -38,11 +38,6 @@ class PopUpViewController: UIViewController {
         }
     }
     
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(animated)
-        setupPopUpView()
-    }
-    
     override func viewWillDisappear(_ animated: Bool) {
         super.viewWillDisappear(animated)
         
@@ -50,16 +45,6 @@ class PopUpViewController: UIViewController {
             self.dimmedView.alpha = 0
         } completion: { _ in
             self.dimmedView.removeFromSuperview()
-        }
-    }
-    
-    private func setupPopUpView() {
-        view.addSubview(popUpView)
-        
-        popUpView.snp.makeConstraints {
-            $0.centerX.centerY.equalToSuperview()
-            $0.height.equalTo(117)
-            $0.width.equalTo(293)
         }
     }
 }
